% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_bounds.R
\name{print.gamma_bounds}
\alias{print.gamma_bounds}
\title{Print Method for gamma_bounds Objects}
\usage{
\method{print}{gamma_bounds}(x, ...)
}
\arguments{
\item{x}{An object of class \code{gamma_bounds}.}

\item{...}{Additional arguments (not used).}
}
\value{
\code{x}, invisibly.
}
\description{
Print Method for gamma_bounds Objects
}
