% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_rho.R
\name{print.infer_rho}
\alias{print.infer_rho}
\title{Print Method for infer_rho Objects}
\usage{
\method{print}{infer_rho}(x, ...)
}
\arguments{
\item{x}{An object of class \code{infer_rho}.}

\item{...}{Additional arguments (not used).}
}
\value{
\code{x}, invisibly.
}
\description{
Print Method for infer_rho Objects
}
