% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_bounds.R
\name{print.latent_bounds}
\alias{print.latent_bounds}
\title{Print Method for latent_bounds Objects}
\usage{
\method{print}{latent_bounds}(x, ...)
}
\arguments{
\item{x}{An object of class \code{latent_bounds}.}

\item{...}{Additional arguments (not used).}
}
\value{
\code{x}, invisibly.
}
\description{
Print Method for latent_bounds Objects
}
