% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_cc.R
\name{print.power_cc}
\alias{print.power_cc}
\title{Print Method for power_cc Objects}
\usage{
\method{print}{power_cc}(x, ...)
}
\arguments{
\item{x}{An object of class \code{power_cc}.}

\item{...}{Additional arguments (not used).}
}
\value{
\code{x}, invisibly.
}
\description{
Print Method for power_cc Objects
}
