% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_nri.R
\name{print.power_nri}
\alias{print.power_nri}
\title{Print Method for power_nri Objects}
\usage{
\method{print}{power_nri}(x, ...)
}
\arguments{
\item{x}{An object of class \code{power_nri}.}

\item{...}{Additional arguments (not used).}
}
\value{
\code{x}, invisibly.
}
\description{
Print Method for power_nri Objects
}
