% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho_bounds.R
\name{print.rho_bounds}
\alias{print.rho_bounds}
\title{Print Method for rho_bounds Objects}
\usage{
\method{print}{rho_bounds}(x, ...)
}
\arguments{
\item{x}{An object of class \code{rho_bounds}.}

\item{...}{Additional arguments (not used).}
}
\value{
\code{x}, invisibly.
}
\description{
Print Method for rho_bounds Objects
}
