% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_cc.R
\name{print.sample_size_cc}
\alias{print.sample_size_cc}
\title{Print Method for sample_size_cc Objects}
\usage{
\method{print}{sample_size_cc}(x, ...)
}
\arguments{
\item{x}{An object of class \code{sample_size_cc}.}

\item{...}{Additional arguments (not used).}
}
\value{
\code{x}, invisibly.
}
\description{
Print Method for sample_size_cc Objects
}
