% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_nri.R
\name{print.sample_size_nri}
\alias{print.sample_size_nri}
\title{Print Method for sample_size_nri Objects}
\usage{
\method{print}{sample_size_nri}(x, ...)
}
\arguments{
\item{x}{An object of class \code{sample_size_nri}.}

\item{...}{Additional arguments (not used).}
}
\value{
\code{x}, invisibly.
}
\description{
Print Method for sample_size_nri Objects
}
